+++ src/clamp.c
@@ -204,10 +204,12 @@ iterable : iterable\n\
     Clamp the values from this `iterable`.\n\
 \n\
 low : any type, optional\n\
-    The lower bound for clamp.\n\
+    The lower bound for clamp. If not given or ``None`` there is no lower \n\
+    bound.\n\
 \n\
 high : any type, optional\n\
-    The upper bound for clamp.\n\
+    The upper bound for clamp. If not given or ``None`` there is no upper \n\
+    bound.\n\
 \n\
 inclusive : bool, optional\n\
     If ``True`` also remove values that are equal to `low` and `high`.\n\