+++ src/deepflatten.c
@@ -27,17 +27,23 @@ deepflatten_new(PyTypeObject *type,
     PyIUObject_DeepFlatten *self;
 
     PyObject *iterable;
-    PyObject *iterator=NULL;
-    PyObject *iteratorlist=NULL;
-    PyObject *types=NULL;
-    PyObject *ignore=NULL;
-    Py_ssize_t depth=-1;
+    PyObject *iterator = NULL;
+    PyObject *iteratorlist = NULL;
+    PyObject *types = NULL;
+    PyObject *ignore = NULL;
+    Py_ssize_t depth = -1;
     Py_ssize_t i;
 
     /* Parse arguments */
     if (!PyArg_ParseTupleAndKeywords(args, kwargs, "O|nOO:deepflatten", kwlist,
                                      &iterable, &depth, &types, &ignore)) {
         goto Fail;
+    }
+    if (types == Py_None) {
+        types = NULL;
+    }
+    if (ignore == Py_None) {
+        ignore = NULL;
     }
 
     /* Create and fill struct */
@@ -165,13 +171,12 @@ deepflatten_next(PyIUObject_DeepFlatten *self)
 
         /* First check if the item is an instance of the ignored types, if
            it is, then simply return it. */
-        } else if (self->ignore && self->ignore != Py_None &&
-                PyObject_IsInstance(item, self->ignore)) {
+        } else if (self->ignore && PyObject_IsInstance(item, self->ignore)) {
             return item;
 
         /* If types is given then check if it's an instance thereof and if
            so replace activeiterator, otherwise return the item. */
-        } else if (self->types && self->types != Py_None) {
+        } else if (self->types) {
             if (PyObject_IsInstance(item, self->types)) {
                 /* Check if it's a builtin-string-type and if so set
                    "isstring". Check for the exact type because sub types might